refactor(svelte2tsx): 通常属性の整形を attributes/attribute.rs へ切り出す
format_attribute_node / format_attribute_node_segments と数値属性判定・
属性名の case 変換・先行コメントの seg 化を純粋移動。
上流の Attribute.ts に対応させる。出力・ロジックの変更なし。
